Crud of Exam's Result Created

+++ src/StudentManagement.Business/DTOs/ExamResultDTOs/PostExamResultDTO.cs
@@ -11,6 +11,6 @@ namespace StudentManagement.Business.DTOs.ExamResultDTOs
     {
         public Guid StudentId { get; set; }
         public Guid ExamId { get; set; }
-        public byte? Score { get; set; }
+        public int? Score { get; set; }
     }
 }
